requirements.txt
aiosmtplib
argon2-cffi
asyncpg
cerberus
itsdangerous
pyyaml
tornado
